requirements.txt
asgiref==3.2.7
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
distro==1.5.0
Django==3.0.7
django-cors-headers==3.5.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docx2pdf==0.1.7
drf-yasg==1.17.1
facebook-sdk==3.1.0
Faker==4.1.0
fire==0.3.1
future==0.18.2
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
img2pdf==0.4.0
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
numpy==1.19.5
oauthlib==3.1.0
openai==0.2.6
packaging==20.4
pdf2docx==0.5.1
pdf2image==1.14.0
pdftotext==2.1.5
pikepdf==2.4.0
Pillow==8.1.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
PyJWT==1.7.1
PyMuPDF==1.18.6
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
python-docx==0.8.10
python-twitter==3.5
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
tabula-py==2.2.0
termcolor==1.1.0
text-unidecode==1.3
tika==1.24
toml==0.10.1
tqdm==4.56.0
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.2.0
XlsxWriter==1.3.7
